Update patch set 21
Patch Set 21:

For me that's sounds great.
If there would be something to change, we can do this later.

For the issue, I think that everything Craig wanted to have fixed is fixed, so if you also think the same, we could start merging from monasca-common as that one is needed by api and thresh in Java.

thx

Patch-set: 21
Label: Verified=0