Update patch set 1
Patch Set 1:

Yes, size is limited in the requirements we have and basically we cannot allow to loose any log, instead we decide to cut as much as needed. 

The problem I had was in producer though, so I was surprised that you've mentioned consumers as well. Anyway, even if this change goes abandoned, I think I need to extend my Kafka knowledge because I was not aware of that.

Patch-set: 1