Update patch set 4
Patch Set 4:

Do you know if this was passing the Monasca Tempest tests by running locally?

Just wondering if a specific Tempest test should be written to address this issue and if you are running the Tempest tests against the Hibernate code. 

Also, I don't if anyone has enough time, but I thought I would mention that getting the Hibernate code tested by OpenStack CI along with the existing MySQL code, might be a good idea to catch these types of issues. Multiple gates could be setup to run against the Monasca API. One for the current MySQL implementation and one for Hibernate.

I'm going to +1 and leave it to your discretion on when to merge, but it all looks good to me.

Patch-set: 4
Label: Verified=0