Update patch set 1
Patch Set 1:

Hi Ryan
I think id is better than number.
Because id is included in a element but number is not included in one.
So Horizon can indicate easily offset when offset is id.
I saw other openstack projects. Nova and Glance have id offset which is called marker.
So I want to change offset in AlarmDefs and Alarms to id.

Patch-set: 1