Fix alarm definition tests
Most of the Java API alarm tests were randomly failing because the
metrics that should create an Alarm were arriving at the Threshold Engine
before the new Alarm Definition does. Added a small sleep to ensure the
new Alarm Definition gets there first.

The first Alarm test was often failing with the Java API because
keystone was taking to long to respond. Increase the time the Java
API will wait for a response to 5 seconds.

The test: test_list_statistics_with_offset_limit(), previously did not
handle a combination offset of an id and a date being returned. Also,
it did not handle the offset being urlencoded. Fixed both of those
issues.

Change-Id: I9df8d86ff0674066266cc928843d15f83e688dd1