Fix alarm definition tests
Some of the Java API alarm tests were failing because the metrics
that should create an Alarm are arriving at the Threshold Engine
before the new Alarm Definition does. Add a small sleep to ensure the
new Alarm Definition gets there first.

The first Alarm test was often failing with the Java API because
keystone was taking to long to respond. Increase the time the Java
API will wait for a response to 5 seconds.

Change-Id: I9df8d86ff0674066266cc928843d15f83e688dd1