DNM Use tempest-plugin service client registration
Tempest is going to expose a new optional interface in the plugin
class, to automatically register service clients implemented in
a plugin. Along with this the former client manager, renamed to
ServiceClients, is going to move to a stable interface in
tempest.lib.

This is a WIP change to test the interface on Tempest side along
with a real life plugin.

Change-Id: Iad1ba3e9498ee2263e497c079d1524fa7ad12439
Depends-on: I2cefbeec66dfdd25a2ae90554904b3c315f0b457