Fix Tempest Statistics test
Fix monasca_tempest_tests.tests.api.test_statistics.TestStatistics.
test_list_statistics_with_offset_limit by assuming offset is an
opaque value and making sure it is urldecoded and urlencoded as
it should be. Also, previously, the test was trying to use an offset
returned by measurement-list and apply it to metric-statistics.
That is not required to work by the API definition and should not
be attempted

Change-Id: Ib4ad5b7f0450dd9415496276b38bd12f2921e5d7
Depends-On: I2a4ae83cc3abafeb618aea0d5ad28620b9f257a0