Update patch set 12
Patch Set 12:

I guess I would prefer keeping the notification_method_type table as well as the foreign key constraint. The Notification Engine and/or API could on start-up, could just insert into the table all the types that have been specified. Since the type is a primary key, if the key already exist in the table, the insert will fail, which is acceptable. I don't see a reason to ever delete types from this table. The table will continue to grow as types are added, but if a plugin is disable, that type will continue to exist in the DB and cause no problems. When the Notification Engine runs, my assumption is that the type will not be recognized and ignored, after and error is logged.

Patch-set: 12
Label: Verified=0