Update patch set 12
Patch Set 12:

> Just thought about this some more, I think it is the Notification
 > Engine that should on start-up insert the supported types. The API
 > doesn't need to. That implies that the Notification Engine should
 > be configured with new notification types/plugins first.

I like that idea, that way most likely we could drop validNotificationTypes written in configuration in API, right ? and make notification engine govern them + still allow API to query them ? Is that something close to what you had in mind ?

But that there's still one problem on the other side, what if someone wishes to disable certain notification methods after some time - regardless of the reason. I don't think removing notification method is an option here, so it kind of starts to look like we would need to have boolean column in notification_method_type that would indicate if given type is disabled or not ? But maybe possibility of such situation is very low and it could be part of manual administrator action or be implemented in the future. 

Do you find that a potential issue ?

Patch-set: 12