Update patch set 12
Patch Set 12:

I think just having the Notification Engine insert all the types that have been enabled is sufficient. If the Notification Engine is re-configured and types are disabled, I don't see any harm in keeping them.

However, as an improvement, I like the idea of an "enabled" column that you mention Tomasz, being added, and the API querying the list of enabled types on start-up. I don't necessarily see that as a requirement for this review, and personally I would rather avoid a SQL schema change right now, but that is something to consider.

Patch-set: 12