Update patch set 10
Patch Set 10: Code-Review-1

Sorry, I missed this earlier, but in all our responses that return objects, we return a JSON object which contains a "links" and "elements" field. This was originally done for a couple of reasons, such as support for pagination, as well as potential extensibility. I would prefer wrapping the response in a JSON object under elements. I don't think we need to add support for pagination yet, but I could see a case in the future where this might be useful, as well as just being consistent with everything else that is there already.

Patch-set: 10
Label: Code-Review=-1