Update patch set 5
Patch Set 5:

So, the gate is reporting the following:

2016-07-25 19:30:13.713664 | 2016-07-25 19:30:13.713 | {0} setUpClass (monasca_tempest_tests.tests.api.test_monasca_ovs_plugin.TestOvsPlugin) ... SKIPPED: Neutron support is required

so, the test is ignored. If you want this to run in the gate, then I'm assuming that Neutron needs to added in the DevStack configuration. Is that something that you are planning on adding in this review?

Patch-set: 5
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0