Fix Influx statistics and Alarm History behavior
Fix influx statistics query to include the timestamp offset

Fix monasca_tempest_tests.tests.api.test_statistics.TestStatistics.test_list_statistics_with_offset_limit
by assuming offset is an opaque value and making sure it is urldecoded and urlencoded as it should be.
Also, previously, the test was trying to use an offset returned by measurement-list and apply it to
metric-statistics. That is not required to work by the API definition and should not be attempted

Fix the Alarm State tests by using the offset returned from the API and not trying to build one. Also,
ensure offset is urldecoded and urlencoded as should be

Change-Id: Iedccae6584e01c56cc05dcc45f9b124cb7d61109