Update patch set 5
Patch Set 5:

This change is basically ready from my POV, so I will soon remove the DNM from the commit message. There are a couple of failures in the java nv job, they do not seem related to my change but I'd like some confirmation on that from someone in the monasca team.

Patch-set: 5