Update patch set 12
Patch Set 12: Code-Review-1

(2 comments)

Kaiyan -- need to fix anchors in the doc, as well as generate an ID that is consistent across requests when paginating.  See how we did this in the dimensions/values API.  Needed for grafana to piece things back together.

But I tested the vertica implementation -- functions as I'd expect except for that.  Thanks!

Patch-set: 12
Label: Code-Review=-1