Update patch set 8
Patch Set 8:

My concern was that we are trying to keep the ORM stuff working, but there wasn't a devstack configuration that made development and test reasonably easy. So, the end goal is to have a fully functional devstack with Postgres and ORM with a job that runs the tempest tests. However, steps toward that goal are definitely worth it

Patch-set: 8