Update patch set 30
Patch Set 30:

Sorry, but this has slipped past me. As Kaiyan has just pointed out to me, the id's should not be necessary. The metric names, dimension names and dimension values resource all return a unique/distinct array of elements. The name that is returned uniquely identifies the element. Therefore an id is not necessary. 

The metric names resource does not return a name. The dimensions names resource that is being added here, should not return one. And, the dimension values resource, that was already merged, should have the id removed.

Patch-set: 30
Label: Verified=0