Update patch set 30
Patch Set 30:

I don't understand all the internals of Grafana, so I'll need some more details on that. In general, I don't like doing things that are specific to a specific application. We haven't done that anywhere else in the API. But, if Grafana needs it implies that other apps will need it too, then I'll be OK with it. 

After reviewing this some more though, I realized a few things. The monasca names resource returns a 1-D array/list. It is easy to paginate through that list. Each name is distinct, and the name can be used as an offset.

Looking at the dimensions names and values resources some more, it is a 2-D structure. It is 2-D because the metric name is optional. In this sense it is similar to measurements and statistics. I'll need to review some more, but just wanted to make sure that if limit is exceeded, which is 10,000 by default, that the offset returned (the offset in the links.next), is a 2-D offset composed of id and either the dimension name or dimension value.

Patch-set: 30