Update patch set 30
Patch Set 30:

The dimension values (as well as this patch) returned an ID as well as the elements array for consistency with clients -- grafana being the most obvious one.  If we hadn't done it this way, client code would have to be special cased for this API, different from the other ones that paginate results.

Patch-set: 30