Update patch set 32
Patch Set 32:

I had an email exchange on the side, that I wanted to add to this review as Craig and others might be interested.

In the case of dimension values, the metric name is optional. The response for either case is that only a single element is returned in the elements list. 
If only a single element is being returned why wouldn't you just return them "values" field in the "elements"? There doesn't appear to be a reason to have a list of lists structure.

The JSON object that is returned in elements implies that there are multiple elements that can be returned, but that doesn't appear to be the case. If metrics names is not supplied, a single list is returned, and the metric_name is not in the response body.

If you did the above, then you won't need an id, and pagination would work exactly as it did in the case of the metric names resource. Each returned metric name, dimension name, or dimension value is distinct, and only a 1-D array should be returned.
If you hit the "limit" of a query, the offset that is returned in the "next" link is only the dimension value. That only works because you are assuming that the elements contains one element. That isn't what the spec says and as a result, that is a bug. Elements is a list, and you need to offset into potentially multiple elements correctly.

Given my comments above, I think what should happen is that the dimension values are returned in the elements field, similar to the metrics names resource. ID would be eliminated in that process.

Patch-set: 32
Label: Verified=0