Update patch set 6
Patch Set 6:

Hi, I was thinking on this change a bit, and basically we came up with same idea for log-api, however we used handler that rolls the file. 

I like the idea of seperate logging file, it's more openstack IMHO. But I think that it is safer to assume logger that'd archive file after exceeding specific size (for instance 1 GB).

Patch-set: 6