Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > Documentation needs to updated.

Line 15: Sorry, my bad. I'll fix it.

While I was trying to find a proper place to introduce info about sporadic in the documentation, I had a doubt about it: Should sporadic be part of the metric identification? I'm asking it because the way that it is done now, sporadic is just a flag and nothing more and after the first time that the metric is published using sporadic=True, this metric will never be non-sporadic again.

If sporadic is part of the metric identity, we can have two different metrics with the same name and dimensions, but one sporadic and the other non-sporadic. This way makes more sense to me.

What do you think about it?

Patch-set: 4
Reviewer: Gerrit User 19861 <19861@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0