Update patch set 4
Patch Set 4:

> > (1 comment)
 > >
 > > Documentation needs to updated.
 > 
 > Line 15: Sorry, my bad. I'll fix it.
 > 
 > While I was trying to find a proper place to introduce info about
 > sporadic in the documentation, I had a doubt about it: Should
 > sporadic be part of the metric identification? I'm asking it
 > because the way that it is done now, sporadic is just a flag and
 > nothing more and after the first time that the metric is published
 > using sporadic=True, this metric will never be non-sporadic again.
 > 
 > If sporadic is part of the metric identity, we can have two
 > different metrics with the same name and dimensions, but one
 > sporadic and the other non-sporadic. This way makes more sense to
 > me.
 > 
 > What do you think about it?

Logically I'd expect that metric with given name is uniquely identified as sporadic or periodic. @Roland, would that be right assumption ? I believe so, but that's a chunk of functionality, so better to establish some stuff to be sure.

Another question is, are we permitted to commit this now, or is still some sort of code-freeze phase ?

Patch-set: 4