Update patch set 3
Patch Set 3:

> Tomasz, I've reached out to Ryan Bak to see if he can take a look.
 > Another option is to temporarily disable Grafana in the build to
 > get CI up and running again. --Roland

Thx Roland, that's kind of bad timing for grafana to brake. Would be great if Ryan could take a look. I've been playing out with setting this up, but either ended up with the same error we see at the gate or the one telling that it is impossible to checkout specific rev/commit from marcano repository (I believe it was the name of it).

Patch-set: 3