Update patch set 5
Patch Set 5:

> (1 comment)

What is more, devstack plugin is kind of broken. I mean not functionally, but big tarballs (like storm, kafka) are downloaded for each stack operation.
I know this is not the case for this change, but it's good to know it happens to work like this. 

I think I even submitted an enhancement to devstack @ launchpad about that fact.

Patch-set: 5