Update patch set 4
Patch Set 4:

Sorry, Ryan

I spent some time trying to understand this change and how statistics works but work now I just could not comprehend that.

I understand that the case is to exclude the next bucket from the result list. 
So assuming default 300s bucket, and start of the bucket:

  2016-10-10T00:00:00.000Z
 
if measurement timestamp is 

  2016-10-10T00:05:00.000Z

this measurement should be included in the bucket, therefore the entire bucket, because adding 5 minutes and removing 1 millisecond gives us
 
  2016-10-10T00:04:59.999Z

right.

This is the theoretical background for this change ?

As far as it goes for the testing, I couldn't come up with anything useful that would prove if your change works or not by hand :(.

Patch-set: 4