Update patch set 2
Patch Set 2: Workflow-1

Couple of typos that I will try and remember to fix later, but let's check if that fix is passing.

That is a bit raw change, I am testing it only as I write down - just wanted to push it here to let guys at least see it and maybe test.

Patch-set: 2
Label: Workflow=-1