Update patch set 1
Patch Set 1:

The only question left here to answer is:

Do we want to enforce

  -- depth 1

for twc-openstack/grafana repo.

Unfortunately I could not add this project to be downloaded using git_clone wrapper (to benefit from GIT_DEPTH variable) in my other change because that requires adding it to PROJECTS variable in gate setup and since grafana is not pure monasca project using plain git clone is the only option.
However without it GIT_DEPTH cannot be directly used. 

But the thing about this particular repository is that it is very big and cloning it takes some time which can be saved if we just clone recent commit. That would speed up also gate setup BTW, not only local deployment time.

What do you think ?

Patch-set: 1