Update patch set 3
Patch Set 3:

Hi Shinya, Thanks again for adding these features. You missed the last Monasca/Cassandra discussion that we held on Monday. Over the past 4 weeks the HPE team has done an in-depth analysis of Cassandra. Unfortunately, we've been unable to resolve all the issues that we've encounters. I've added some of the comments in the etherpad at, https://etherpad.openstack.org/p/monasca_cassandra. Not sure that the list will be clear without further explanation. Please see the issues section starting at 6.2. Sorry to be the bearer of bad news, but it isn't clear that Cassandra is going to be an acceptable solution for Monasca.

Since this code is ready to go, we can merge it and I guess this will make the Cassandra implementation functionally complete. But, the overall implementation is not going to be performant.

Patch-set: 3