Update patch set 9
Patch Set 9: Code-Review+1

Tested with workaround. Everything peachy :D

For the workaround, I think we may try with updated documentation.
Not sure if putting initial state of the VM with SATA set into the snapshot would work. 

Anyway, everything worked for me, so we should move with trying to get this and the infra change merged

Patch-set: 9
Label: Code-Review=+1