Removed unnecessary sleep from devstack plugin
The step installing influxdb sleeps for a full minute before allowing the
plugin to progress.  The commented explanation is to wait for influxdb to
start, elect a leader and begin accepting connections.  However, the
subsequent part of the process is to install various packages (via apt and git)
then build the java or python monasca-api code.  The chances of that being done
before a minute is out are low.  The sleep is therefore redundant and
aggravating.

Change-Id: I985bede38890a0ab0abafc93bda0feaafe2fa99b