Update patch set 1
Patch Set 1: Code-Review-2

Hi Adam, We can't merge this change into the Newton branch as it is not a critical bug fix. We can only fix bugs in official releases. 

We discussed in the monasca weekly team meeting what we can do today. The problem is that the monasca-grafana-datasource doesn't have an official release with Newton. After the Newton release we added a couple of new endpoints to the Monasca API for supporting dimensions queries. The monasca-grafana-datasource was modified to use them. However, since the monasca-grafana-datasource isn't tagged/releases for Newton, when you pull the Monasca Newton release, the monasca-grafana-datasource is now incompatible. YOu could potentially pull a specific commit from the monasca-grafana-datasource to resolve the problem.

We are in the process of trying to tag the monasca-grafana-datasource for the Newton and Ocata release. For the Newton release we will try and add the monasca-grafana-datasource with a 1.0.0 release tag and a compatible commit. For the Ocata release we will try and add the monasca-grafana-datasource with a 2.0.0 release tag.

If you have any further questions please let me know.

--Roland

Patch-set: 1
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2