Add support for InfluxDB 1.0.2
Somewhere between InfluxDB version 0.9.3 and 1.0.2 InfluxDB changed the
response object returned when querying series. The "name" field was
removed from the values, and the values was changed to a comma
separated string where the first element in the string is the metric
name and the remaining elements are the tags. Modifed the code to be
compatible with either the old or new response object. Hopefully, there
weren't any intermediate changes.

Change-Id: I33188f4e64150d2cb24c0a06979c06f03c803da4