Update patch set 3
Patch Set 3:

Personally, I think this is a good change in right direction. Several different projects in openstack are using oslo.db to handle this. However the scope of migrating to oslo.db is much broader and includes also changes in:
- configuration
- repository

Are you planning on including all that ?

@Roland, what's your opinion on this ?

Patch-set: 3
Label: Verified=0