Update patch set 14
Patch Set 14:

Tomasz, With respect to the question regarding Cassandra, I think the answer is yes, as the performance isn't that competitive to other solutions, such as InfluxDB or Vertica. There is no one else that wishes the answer was different, but that is our honest opinion based on the analysis we've done, which has been extensive. Cassandra would provide the scale out architecture and reliable distributed data. However, the starting point in performance is not so good, and ultimately the queries will be very difficult to add, and due to issues with the effective row size limits which are around 100K elements, which lead to further partitioning the data which in-turn makes it very hard to search for metrics based on name, dimensions and timestamp, as well as removing the possibility for in-database user-defined functions and aggregates, and the requirement to still add compression, I don't think Cassandra is going to result in a answer that we will be happy with, unless your only goal was a scale-out distributed reliable datastore.

It should also be noted that InfluxDB did their own analysis a few months ago, and granted they have their own business interests and might not be viewed as unbiased, I think what they reported on was reasonably accurate. In fact, we set out to prove them wrong in a sense, and I think we just validated their analysis the more we got into it.

So, I still don't rule out Cassandra yet. I'm still hoping for a miracle or aha moment. Maybe Cassandra is still good enough in performance for your needs, and provides the data reliability/consistency you need. Which is where the open-source version of InfluxDB has problems.

Patch-set: 14