Update patch set 7
Patch Set 7:

Hmm...looks like using sudo to set up services messes around with run_process function. This is more than I anticipated there would be to do. I am not sure if I will be able to work on this right now. However the entire idea behind this change remains as it was - be able to run monasca not just only on ubuntu

Patch-set: 7