Update patch set 3
Patch Set 3: Workflow-1

This one needs to sit around because first we need to address the problem that monasca-api plugin does not follow the standard other plugins do. At least  this is how it looks at this point.

Once postgres change is merged which indirectly address that issue is merge I will get back to it.

Patch-set: 3
Label: Workflow=-1