Update patch set 1
Patch Set 1:

Hi Tomasz, This looks reasonable to me. Now that we have the Tempest tests, are the smoke tests serving any inherent need that the Tempest tests can't serve? The smoke tests are possibly a little more specific and aren't just for API tests, but I'm not sure that they are needed anymore and would clean-up the Monasca DevStack plugin a bit.

Patch-set: 1