Update patch set 18
Patch Set 18: Code-Review+1

I think this is ready to go. I will look into optimizing the InfluxDB queries to use SHOW TAG KEYS and SHOW TAG VALUES in a subsequent review, but I wanted to first get support for InfluxDB in the Python code up to the latest InfluxDB release, and then look into changing the queries to be possibly more optimal, which is a bigger job.

Patch-set: 18
Label: Code-Review=+1