Update patch set 18
Patch Set 18:

Tomasz, I thought about your comments on adding unit tests. In this situation I don't think the unit tests are really enhancing our quality so I don't have any plans to add any. The Tempest Tests are to a large extent already testing the "unit" level functionality and the unit tests are IMHO testing basic things, that aren't adding anything to the Tempest Tests, and therefore would be redundant. Case in point, the unit tests found no problems with the change from InfluxDB 0.9.5 to 1.1, and were not used to develop any of the the new code. If you look at the current set of unit tests, all they do is check the happy case of invoking the function given what we think the InfluxDB data structure looks like, which doesn't necessarily reflect reality, and validating that the code that is in the function processes the happy case, and returns a structure that it expect. We already test that case thoroughly in the Tempest tests, and it is done against a live InfluxDB server. In addition, we test many of the not happy cases in the Tempest tests.

Patch-set: 18