Update patch set 2
Patch Set 2:

Hmm.  I'm not so sure about this.  I'm having difficulty getting the devstack environment to work but it may be to do with horizon.  If I disable horizon it all seems to work.  But with horizon it fails and the devstack filesystem ends up readonly!

I am bewildered.

Patch-set: 2
Label: Verified=0