Update patch set 7
Patch Set 7:

> I don't think we should be removing the KAFKA_VERSION information
 > from our settings.  By default the devstack kafka plugin is
 > installing a version of Kafka we haven't validated.

KAFKA_VERSION is saved in override_defaults file. This is because kafka-plugin specifies that on its own and we have to override (as the name suggests). At least this is what I initially thought, so I will just try this once again and see if this needs to stay in override_defaults or can be moved back to the settings

Patch-set: 7
Label: Verified=0