Update patch set 1
Patch Set 1: Code-Review+1

Hi Shinya, I tested this and it worked fine. The tempest tests don't pass, but one of the failures is a bad test and I haven't had time to debug the other two, but they don't looked related to using Vertica 8.0. I agree with your logic about using 8.0 as I also saw it was the only community version available

Patch-set: 1