Update patch set 9
Patch Set 9:

> Tomasz, looking at the gate job log I can't see that it did install
 > the proper version of Kafka.  It doesn't appear to have $KAFKA_DEST
 > set though so the kafka commands to create topics failed.

That's because this is the plugin, the case is to test this manually and if someone confirm this works we can create NV gate to verify this at CI level. Or I can add this gate right now, either way it is fine with me.

Patch-set: 9
Label: Verified=0