Update patch set 2
Patch Set 2:

Please ignore my previous comment about setting -Djava.net.preferIPv4Stack - I've not tested that.

All in all, I've a problem with tempest tests for monasca-log-api, and after some investigation I've found out that some services (in my case this is monasca-log-transformer, which is a logstash script) are unable to connect to Kafka, if the host machine has IPv6 address assigned. 

Currently Kafka is bound to listen only on an external interface, and I'd like to have it listen on all available interfaces. This resolves my problem with IPv4 vs. IPv6 addressing

Patch-set: 2