Modify Storm and Kafka download url and Vagrant box
Modify Vagrant box to bento/ubuntu-16.04. This resolves the issue
with the read only file system due to the scsi driver used in
ubuntu/xenial64. bento/ubuntu-16.04 uses the SATA driver and the
prior issues with using bento have been resolved. Note, we had
tested bento earlier and it did not work, but whatever issue that
we had been running into has been resolved.

Modify the download url for Apache Storm and Apache Kafka from
tds to nextag or nexcess. tds was extremely slow. nextag and
nexcess was the fastest that I tested.

Modify the Threshold Engine from version 2.0.0 to 2.1.0 as it is
currently not installable with 2.0.0.

Change-Id: Ia754dc83fe65f391bee1e5e51851334a856309e8