Update patch set 6
Patch Set 6:

The database is chosen at install time, and potentially changed later, so it's up to whoever is installing/configuring Monasca to choose the right package to make influx or cassandra work.  I don't know how you'd like to do that via tox but I don't think adding it to the requirements is the right approach.

Patch-set: 6