Add prometheus instrumentation and /metrics endpoint
This adds a new /metrics endpoint (not to be confused with
/v2.0/metrics) that reports Prometheus-compatible metrics using the
python prometheus_client package.

Due to the nature of the Prometheus endpoint detection in
monasca-agent, the new metrics endpoint must a) be exposed by the
API process directly on the same port and b) must not require keystone
authentication. As such, this patch also requires
`delay_auth_decision` in the keystone_authtoken config to be enabled,
and adds a new `skip_keystone_auth` property that can be set in Falcon
handlers (read by the keystone_context_filter).

As monasca-api is generally run with multiple processes under a WSGI
handler like gunicorn, an additional environment variable must also
be specified, `prometheus_multiproc_dir`, to be used for sharing state
between the separate processes. This environment variable can be set
to any empty, temporary directory. Additional parameters have also
been added to api-config.ini to handle cleanup of these data files
during gunicorn lifecycle events.

Change-Id: Ia5f76177a2e34e1703b48407aeabe011657b646a