Add prometheus instrumentation
This adds instrumentation using the python prometheus client that
sends metrics to a monasca-sidecar [1] instance for eventual scraping
by the monasca agent's prometheus plugin.

The use of a sidecar server resolves issues arising from normal
prometheus instrumentation in an application with multiple processes.
This allows liveness detection and client-side aggregation to be
performed with minimal invasive modifications to the host app, i.e.
monasca-api.

[1] http://github.com/timothyb89/monasca-sidecar

Change-Id: Ia5f76177a2e34e1703b48407aeabe011657b646a