Fix influxdb statistics list with multiple functions
If multiple statistic functions were given, for example, max,min, there
would be too many rows returned.

The fix is to only append the row once

The test that shows this is fixed is in https://review.openstack.org/417180/.
It is in a separate review because I was fixing the test to work with Vertica
and the corrected test showed up this bug for Influxdb implementation.

Change-Id: Ifbb10f42d7849eb9b36f6bf1f8f8f7fef2940e2e
Closes-Bug: #1656122