Update patch set 8
Patch Set 8:

@Tim, is this monasca-sidecar meant to replace monasca-statsd ?
Just asking :), because if monasca-log-api is based on monasca-statsd to collect internal metrics (as well as monasca-notification) and if monasca-sidecar will be used instead it sure look like (to me) as some sort of conflict. I mean I am not saying I don't like it, but would like to know some more details.

Patch-set: 8
Label: Verified=0