All alarm-definition-updated event fields must be set
This fixes a very subtle bug I discovered testing last week. On an
alarm definition patch, not all of the alarm-definition-event fields
were being set, for example, actionsEnabled. Since that was set to None,
the Threshold Engine read the message as setting actionsEnabled to false
on the Alarm Definition and then when an Alarm went to the ALARM state,
thresh set actionsEnabled to false in the AlarmStateTransitionEvent message.
This caused the Notification Engine to not issue a notification for that
AlarmStateTransitionEvent, even though it should have. Took me a while
to track this one down.

Added a unit test to ensure this doesn't happen again. I don't know if
using Mock is the right way to get at the send_event arguments, but it
was the only way I could come up with. If someone has a better way, please
let me know

Change-Id: I6822d0f84345d252634a923b2a887b9b4638b6d6
Depends-On: I0de4875300f77ea4219576195af967f944cf8005