Update patch set 2
Patch Set 2:

Well, I think mock is fair enough here but that test is wow :).

I will sit with Artur and use our limited set of skill to review that but in the meantime, would you remove this last paragraph from the commit message and move it to the comment in the change ? Cheers.

Patch-set: 2
Label: Workflow=0