Update patch set 3
Patch Set 3:

Hi @Tomasz and @Atur . Kaiyan and I are trying to fix https://bugs.launchpad.net/monasca/+bug/1659071 in our patchsets. We like the refactor that you have presented in this patchset and if you would like to fix this bug here, we would be happy to use it.
I think the problem we are experiencing with bringing up a local Devstack has to do with the git_clone function versioning improperly. In functions-common, we pass a –depth 1 flag into the git_clone function which causes the library to do a shallow clone with a history truncated and a version number of 0.0.0. Then, when the other libraries come to install, they see a version of 0.0.0 and they try to uninstall and clone the master branch, but can’t because of the sudos. We could easily fix the permissions problem that makes local Devstack fail with removing all appropriate sudos, but that would defeat the purpose of testing any local code.
Please let us know if you would like to modify this patchset to fix that bug. Kaiyan and I are currently trying to use this patchset to see if we can find a solution, so we can also let you know if we figure anything out.

Patch-set: 3