Fix "permission denied" error in devstack
Devstack was failing locally and we think the problem was git_clone
funtion is passing git_depth=1 to monasca-common and monasca-statsd
which caused a shallow clone with missing history. That's why we got
error with monasca-common version 0.0.0 which does not even exist.

With this change, we will do git clone for monasca-common and monasca
-statsd with git_depth=0 and other repos with git_depth=1.

Change-Id: Ied733853b4b22368095df7047a07694775bc78a5
Depends-On: I226e42655c834a9a8ac7e58989ac86018d7d4576
Closes-Bug: #1659071