Update patch set 8
Patch Set 8:

just fyi, I did a lot of work on providing consistent statsd-instrumentation for the Python implementations of Monasca. I am planning to upstream then (they are in github/sapcc fork today) once they have proven in practise.

With the prometheus-statsd exporter, my hopes are that we can stick with statsd instrumentation and this way support additional options for fall-back monitoring.

Patch-set: 8