Update patch set 8
Patch Set 8:

@Tomasz, Joachim: we did run into some challenges while looking at statsd, I attempted to document them on a wiki page here: https://wiki.openstack.org/wiki/Monasca/Instrumentation

There's still a bunch of open questions and plenty of discussion to be had, but the short of it is that we're not sure monasca-statsd meets all our desires, at least in its current form. That said, I don't want to suggest this is necessarily the correct solution either, it's really intended as a proof-of-concept to get some discussions going.

Aside, I've also created blueprints:

https://blueprints.launchpad.net/monasca/+spec/prometheus-instrumentation
https://blueprints.launchpad.net/monasca/+spec/monasca-sidecar

Patch-set: 8