Update patch set 2
Patch Set 2:

Guys, the first problem with bandit encountered here is about using assert in code. That was rather done on purpose, so the question to you is should we just ignore this error or rework the code to use some other structure to check if everything is ok in certain conditions.

Patch-set: 2
Label: Verified=0