Update patch set 9
Patch Set 9:

I think that this change is incomplete.

Having influxdb and cassandra in requirements is good idea because it is controlled by the bot, unline tox.ini + here we should also change versions of either clients installed in devstack environment.

For instance:

  test-requirements [ influx ] >= 2.9.2
  devstack          [ influx ] == 2.8.0

I will try to normalize all that, to have single version of influx wherever we need, but I guess we can do it after stable/ocata, right ?

Patch-set: 9
Label: Verified=0