Remove mysql repositories
The sqla repositores are used for all tests and the mysql repositories aren't
used or tested anymore. Reduce confusion in the code by deleting the dead
mysql reposistories.

Change etc/api-config.conf to use sqla to match the devstack files

Remove the mysql config section, only the database section should
be used now.

Set the repository defaults to be the correct ones in the config file
definition.

Change-Id: I748e91cadb059d91380b78d1e2af0ff0001fc397