Fix Python tempest gate
When monasca-common is installed from source for the API, it does
not follow upper constraints so it is installing PyMySQL-0.7.10.
When the API is installed, it is does a pip_install PyMySQL which
uses the upper constraints so it wants to use PyMySQL-0.7.9. The
install fails uninstalling PyMySQL-0.7.10.

Change the install of monasca-common from source to use upper
constraints.  Change all of the other installs for monasca-common and
monasca-statsd from source to also use upper constraints using the
devstack utility function setup_install

Change-Id: I6535f131bcc1be2ed06b7011da51a6ac887ec2ee