Add API Resource for Alarm Silencing Manager
Silencing manager will mute alarms for a given time based on
input parameters: start_time and end_time. A silence is configured
based on matchers and if an incoming alarm is checked to match the
matchers, no notifications will be sent out for that alarm.

Depends-On: I126e9a8e9fdec9f1f1a7f0c09033c256f4099b54
Change-Id: Ica6148e96cb9fa74ec6488564cca48019af69a02
Partially-Implements: blueprint alarm-manager-resource
Story: 2000935
Task: 4068