Update patch set 4
Patch Set 4:

Yes, I agree that the long term solution would be better. But for now I think we should try to fix this bug by using the same functional implementation as in java. With this small change we will have the "colon in dimension-value" solved. And we already know, that the http_status mon-agent plugin creates such dimension-values. For the long term solution I would propose to create a new enchancement. Because changing the syntax for dimensions will have a much bigger impact on other components like monasca-ui, grafana-monasca-datasource, python-monascaclient, etc. And we also need to consider on the backwards compatibility.

Patch-set: 4