Update patch set 31
Patch Set 31:

Artur, can we also put 

  devstack/files/env.sh

into this change ? It contains the environmental variables we've been talking about. I think that we ought to source openrc back there with mini-mon user&project and just export the password as in

  export OS_PASSWORD="password"

because by default that will point to wrong password.

Patch-set: 31