Update patch set 9
Patch Set 9: Workflow-1

@Tomasz, if the SIDECAR_URL environment variable is unset no metrics will be sent anywhere.

I haven't seen osprofiler but it would be good to look at. If it's easy adapt the SidecarBridge to use it we can drop the dependency on the prometheus client. It looks like osprofiler has some notion of notifier/collector plugins which would be exactly what we're looking for.

Patch-set: 9
Label: Workflow=-1